test: re-arrange into views
